refactor(constants): 상수 특성별 분리
