+++ src/components/add-product-form.ts
@@ -1,5 +1,5 @@
 import Component from '../abstract/component';
-import { ACTION } from '../constants';
+import { ACTION } from '../constatns/flux-constants';
 import { customElement } from '../decorators/decortators';
 import createAction from '../flux/createAction';
 import Store from '../flux/store';
