+++ src/constatns/auth-constants.ts
@@ -1,0 +1,13 @@
+export const MEMBER = {
+  SUCCESS_SIGN_UP: '가입이 완료되었습니다.',
+  PLEASE_EMAIL: '이메일을 입력해주세요.',
+  PLEASE_PASSWORD: '비밀번호를 입력해주세요.',
+  SUCCESS_LOGIN: '로그인 완료',
+  SUCCESS_MODIFY_INFO: '정보가 변경되었습니다.',
+  SUCCESS_LOG_OUT: '로그아웃 되었습니다.',
+
+  DIFFERENT_PASSWORD: '비밀번호가 다릅니다. 다시 확인해주세요.',
+  NEED_MORE_COMPLICATED_PASSWORD: '영문,숫자, 특수문자 중 2가지 이상을 혼합하여 입력해주세요.',
+  NO_REPEATED_CHAR: '같은 문자를 4번 이상 사용하실 수 없습니다.',
+  NO_SPACE_REQUIRED: '비밀번호는 공백 없이 입력해주세요.',
+};
