+++ src/constatns/routing-constants.ts
@@ -1,0 +1,10 @@
+export const PAGE = {
+  LOGIN_INFO: '<login-info></login-info>',
+  LOGIN_HEADER: '<login-header></login-header>',
+  PRODUCT_MANAGE_PAGE: '<product-manage-page></product-manage-page>',
+  CHARGE_MONEY_PAGE: '<charge-money-page></charge-money-page>',
+  PURCHASE_PRODUCT_PAGE: '<purchase-product-page></purchase-product-page>',
+  LOGIN_FORM: '<login-form></login-form>',
+  SIGN_UP_FORM: '<sign-up-form></sign-up-form>',
+  MODIFY_MEMBER_FORM: '<modify-member-form></modify-member-form>',
+};
