+++ src/validation/validation-error.ts
@@ -1,4 +1,4 @@
-import { VALIDATION_ERROR_NAME } from '../constants';
+import { VALIDATION_ERROR_NAME } from '../constatns/validator-constants';
 
 class ValidationError extends Error {
   constructor(message: string) {
